third_party/WebKit/LayoutTests/bluetooth/characteristic/readValue/service-is-removed.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<script src="../../../external/wpt/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
bluetooth_test(() => {
  return setBluetoothFakeAdapter('HeartRateAdapter')
    .then(() => requestDeviceWithTrustedClick({
      filters: [{services: ['heart_rate']}],
      optionalServices: ['generic_access']}))
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService('generic_access'))
    .then(service => service.getCharacteristic('gap.device_name'))
    .then(characteristic => {
      return setBluetoothFakeAdapter('MissingServiceHeartRateAdapter')
        .then(() => assert_promise_rejects_with_message(
          characteristic.readValue(),
          new DOMException(
            'GATT Service no longer exists.',
            'InvalidStateError'),
          'Service got removed.'));
    });
}, 'Service gets removed. Reject with InvalidStateError.');
</script>
